Created isProxy.js
Monkey patch for the native Javascript Proxy constructor function. This adds a __isProxy property to any created Javascript proxies which returns true when accessed. This does not modify the original target object nor is there any other way to detect if a Proxy is being used (unless you want to go with things like the postMessage hack but this is cleaner).